+++ code/setStrings11.cpp
@@ -12,7 +12,7 @@ int main() {
     // cin >> input_string;
     getline(cin, input_string);
 
-    for (int i = 0; i < input_string.length(); i++) {
+    for (size_t i = 0; i < input_string.length(); i++) {
         if(input_string[i] != ' ') {
             set_of_chars.insert(tolower(input_string[i]));
         }
